requirements.txt
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.36.0
google-genai>=0.2.0






gspread
google-auth
